% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGTranscriptome_retrieval.R
\name{connect}
\alias{connect}
\title{Connect to the database}
\usage{
connect(db = "csvdb")
}
\arguments{
\item{db}{database to connect to (defaults to csvdb)}
}
\value{
an sqlite connection
}
\description{
Connects to the SQLite database
}
\examples{
connect(db="csvdb")
}
