% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGTranscriptome_retrieval.R
\name{convertGene}
\alias{convertGene}
\title{Convert gene name to uppercase}
\usage{
convertGene(gene)
}
\arguments{
\item{gene}{string (gene name in upper or lower case)}
}
\value{
string
}
\description{
Convert gen name to uppercase
}
\examples{
convertGene("S100a9") # returns S100A9
}
