% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGTranscriptome_retrieval.R
\name{getExpression}
\alias{getExpression}
\title{Return expression matrix for a given gene}
\usage{
getExpression(connection, dataset, gene)
}
\arguments{
\item{connection}{sqlite connection}

\item{dataset}{string (dataset to retrieve ids from)}

\item{gene}{gene to get expression values for}
}
\value{
data frame of expression values
}
\description{
Return an expression matrix that contains values
for each probe/ensembl id for the given gene
}
\examples{
getExpression(connect(), "MIGTranscriptome_0001", "S100a9")
}
