% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGTranscriptome_retrieval.R
\name{getMatrix}
\alias{getMatrix}
\title{Return matrix of expression values for a given dataset}
\usage{
getMatrix(connection, dataset)
}
\arguments{
\item{connection}{sqlite connection}

\item{dataset}{string (dataset to retrieve matrix for)}
}
\value{
data frame
}
\description{
Return a normalised expression matrix for a given
dataset in the database
}
\examples{
getMatrix(connect(), "MIGTranscriptome_0001")
}
