% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGTranscriptome_retrieval.R
\name{getMetadata}
\alias{getMetadata}
\title{Get metadata}
\usage{
getMetadata(connection, dataset)
}
\arguments{
\item{connection}{sqlite connection}

\item{dataset}{string (dataset to retrieve metadata for)}
}
\description{
Return the metadata for a given dataset
}
\examples{
getMetadata(connect(), "MIGTranscriptome_0001")
}
