% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGTranscriptome_retrieval.R
\name{getProbes}
\alias{getProbes}
\title{Return probe/ensembl id set}
\usage{
getProbes(connection, dataset, gene)
}
\arguments{
\item{connection}{sqlite connection}

\item{dataset}{string (dataset to retrieve ids from)}

\item{gene}{gene to search for matching probes/ensembl ids}
}
\value{
vector of probes/ensembl gene ids
}
\description{
Get all probes/ensembl ids that match a given gene
from the database
}
\examples{
getProbes(connect(), "MIGTranscriptome_0001", "S100a9")
}
