% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGTranscriptome_plotting.R
\name{plotGOI}
\alias{plotGOI}
\title{Plot expression values for a gene of interest}
\usage{
plotGOI(dataset, mat, metadata, variable = "treatment")
}
\arguments{
\item{dataset}{string (dataset to retrieve data from)}

\item{mat}{expression matrix(subsetted)}

\item{metadata}{dataframe of metadata}

\item{variable}{variable present in metadata to colour by}
}
\description{
Return a plot of expression values
}
