% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGTranscriptome_retrieval.R
\name{showDatasets}
\alias{showDatasets}
\title{Show datasets and their attributes}
\usage{
showDatasets(connection)
}
\arguments{
\item{connection}{RSQLite database connection}
}
\value{
data frame of datasets and their attributes
}
\description{
Display datasets and attributes related to those
datasets including metadata and contacts
}
\examples{
showDatasets(connect("csvdb"))
}
