% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGTranscriptome_retrieval.R
\name{sortMetadata}
\alias{sortMetadata}
\title{Sort metadata}
\usage{
sortMetadata(mat, metadata)
}
\arguments{
\item{mat}{expression matrix (data frame)}

\item{metadata}{metadata returned by getMetadata (data frame)}
}
\description{
Sort metadata by sample names given in an expression matrix
}
\examples{
mat <- getMatrix(connect(), "MIGTranscriptome_0001")
metdata <- getMetadata(connect(), "MIGTranscriptome_0001")
sortMetadata(mat, metadata)
}
